Added more systems, refactored some ecs
Modified animation and added sprite components. These are by no means
final, and are just there to give a rough outline of how such systems
might look and work.

Added create_character_model and a resource<model> container in main for
testing the new animations within ecs
ecs systems now take whatever variable arguments context::update(...) is
called with, in the game_context case, this would be a reference to
itself, so render systems can grab the window and draw directly to it.