Added compress_entities and swap_components
Whenever an entity dies, context::is_compressed will be set to false.

After each update, if the context is no longer compressed,
compress_entities will be called and the entities and component arrays
will be sorted by "aliveness", and the is_compressed bool set to true.

This way the update system doesn't have to trudge through lots of dead
entities just to update the live ones.