Added a wall model
Each animation has the wall connecting in different areas, named by whether the direction connects or not. The order is top-right-bottom-left clockwise.

So a wall connects above and to the left, is named "1001", and an intersection is named "1111".

Each animation also has 4 frames for different types of wall included in the spritesheet.